Fixes frontend contribution issues
