+++ app/package.json
@@ -28,6 +28,7 @@
     "connected-react-router": "^6.5.2",
     "css-loader": "2.1.1",
     "date-fns": "^1.30.1",
+    "date-fns-timezone": "^0.1.4",
     "dotenv": "6.2.0",
     "dotenv-expand": "4.2.0",
     "eslint": "^5.16.0",
