+++ app/src/api/api.js
@@ -339,77 +339,6 @@ export const PurposeTypeEnum = Object.freeze({
   UTILITIES: 'utilities',
 });
 
-// TODO: need to fix some of the fields here.
-export const mapContributionFormToData = data => {
-  const {
-    streetAddress,
-    amountOfContribution,
-    city,
-    dateOfContribution,
-    addressLine2,
-    firstName,
-    lastName,
-    entityName,
-    state,
-    zipcode,
-    employerName,
-    employerCity,
-    employerState,
-    occupation,
-    inKindDescription,
-    electionAggregate,
-    email,
-    phone,
-    phoneType,
-    checkNumber,
-    typeOfContributor,
-    subTypeOfContribution,
-    typeOfContribution,
-    inKindType,
-    oaeType,
-    submitForMatch,
-    paymentMethod,
-    notes,
-    isPerson = !!(
-      typeOfContributor === ContributorTypeEnum.INDIVIDUAL ||
-      typeOfContributor === ContributorTypeEnum.FAMILY
-    ),
-  } = data;
-
-  const transformed = {
-    city,
-    firstName: isPerson && firstName ? firstName : null,
-    middleInitial: '',
-    lastName: isPerson && lastName ? lastName : null,
-    name: entityName || null,
-    state,
-    occupation,
-    employerName,
-    employerCity,
-    employerState,
-    checkNumber,
-    contributorType: typeOfContributor,
-    subType: subTypeOfContribution,
-    type: typeOfContribution,
-    inKindType: inKindType || null,
-    oaeType,
-    address1: streetAddress,
-    address2: addressLine2,
-    email,
-    phone,
-    phoneType: phoneType || null,
-    amount: parseFloat(amountOfContribution),
-    date: new Date(dateOfContribution).getTime(),
-    zip: zipcode,
-    inKindDescription,
-    calendarYearAggregate: electionAggregate,
-    submitForMatch: submitForMatch === 'Yes',
-    paymentMethod,
-    notes,
-  };
-  return transformed;
-};
-
 export function post(url, data) {
   const headers = {
     'Content-Type': 'application/json',
